requirements.txt
netifaces
pycryptodome~=3.10
aioconsole